Added and fixed files
