+++ README.md
@@ -1,1 +1,1 @@
-# project_3
+# project-3
