+++ app.py
@@ -1,20 +1,19 @@
-from flask import Flask, render_template, jsonify
+from flask import Flask, render_template, jsonify, send_from_directory
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.ext.automap import automap_base
-
-
+import os
 #Reflect the tables
-engine = create_engine('sqlite:////Users/nguyengiang/project-3/data/Accidents.sqlite')
+engine = create_engine('sqlite:///Data/Accidents.sqlite')
 Base = automap_base()
 Base.prepare(engine, reflect = True)
 
 #Save references to each table
-Accidents2017 = Base.classes.Accidents_2017
-Accidents2018 = Base.classes.Accidents_2018
-Accidents2019 = Base.classes.Accidents_2019
-Accidents2020 = Base.classes.Accidents_2020
+Accident2017 = Base.classes.Accident_2017
+Accident2018 = Base.classes.Accident_2018
+Accident2019 = Base.classes.Accident_2019
+Accident2020 = Base.classes.Accident_2020
 
 app = Flask(__name__)
 
@@ -28,13 +27,13 @@ def get_accidents(year):
     session = Session()
 
     if year == 2017:
-        data = session.query(Accidents2017).all()
+        data = session.query(Accident2017).all()
     elif year == 2018:
-        data = session.query(Accidents2018).all()
+        data = session.query(Accident2018).all()
     elif year == 2019:
-        data = session.query(Accidents2019).all()
+        data = session.query(Accident2019).all()
     elif year == 2020:
-        data = session.query(Accidents2020).all()
+        data = session.query(Accident2020).all()
     else:
         return jsonify({"error":"Invalid year"}), 400
     
@@ -55,6 +54,12 @@ def get_accidents(year):
             "casualty_severity": item.casualty_severity
         })
     return jsonify(results)
+
+# Serve the static GeoJSON files
+@app.route("/geojson/<string:filename>")
+def serve_geojson(filename):
+    geojson_dir = os.path.join(app.root_path, "geojson")
+    return send_from_directory(geojson_dir, filename)
 
 if __name__ == "__main__":
     app.run(debug=True)
