--- css/style.css
@@ -1,28 +1,0 @@
-body {
-    margin: 0;
-    padding: 0;
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .content {
-    flex-grow: 1;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 20px;
-  
-  }
-  
-  #map-container {
-    height: 400px;
-    width: 100%;
-    border: 1px solid black;
-    
-  }
-  
-  #map {
-    width: 400px; 
-    height: 400px; 
-  }
